+++ CHANGELOG.md
@@ -13,6 +13,9 @@ FlatLaf Change Log
 - TabbedPane: Support scrolling tabs with mouse wheel (if `tabLayoutPolicy` is
   `SCROLL_TAB_LAYOUT`). (issue #40)
 - TabbedPane: Repeat scrolling as long as arrow buttons are pressed. (issue #40)
+- TabbedPane: Support adding custom components to left and right sides of tabs
+  area. (set client property `JTabbedPane.leadingComponent` or
+  `JTabbedPane.trailingComponent` to a `java.awt.Component`) (issue #40)
 - Support painting separator line between window title and content (use UI value
   `TitlePane.borderColor`). (issue #184)